# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Resin Trading Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List all resin sources
      parameters:
        - name: resinCategory
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ResinCategory"
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResinEntry"
    post:
      operationId: createSource
      tags: [sources]
      summary: Create a new resin source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResinEntry"
      responses:
        "201":
          description: Created source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResinEntry"
  /sources/{id}:
    put:
      operationId: updateSource
      tags: [sources]
      summary: Update a resin source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResinEntry"
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResinEntry"
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Delete a resin source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /demands:
    get:
      operationId: listDemands
      tags: [demands]
      summary: List all resin demands
      parameters:
        - name: resinCategory
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ResinCategory"
      responses:
        "200":
          description: List of demands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResinEntry"
    post:
      operationId: createDemand
      tags: [demands]
      summary: Create a new resin demand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResinEntry"
      responses:
        "201":
          description: Created demand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResinEntry"
  /demands/{id}:
    put:
      operationId: updateDemand
      tags: [demands]
      summary: Update a resin demand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResinEntry"
      responses:
        "200":
          description: Updated demand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResinEntry"
    delete:
      operationId: deleteDemand
      tags: [demands]
      summary: Delete a resin demand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /matches:
    get:
      operationId: getMatches
      tags: [matches]
      summary: Get potential source-demand matches (paginated)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: resinCategory
          in: query
          required: false
          schema:
            type: string
        - name: entryId
          in: query
          required: false
          description: Filter to matches involving this entry ID
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of matches
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Match"
  /matches/count:
    get:
      operationId: getMatchCount
      tags: [matches]
      summary: Get total match count (fast, uses cache)
      parameters:
        - name: resinCategory
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Match count
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
  /matches/count-by-entry:
    get:
      operationId: getMatchCountByEntry
      tags: [matches]
      summary: Get per-entry match counts for row badges
      parameters:
        - name: resinCategory
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Map of entry ID to match count
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
...
components:
  schemas:
...
    ResinCategory:
      type: string
      enum: [virgin, offgrade, recycled]
    ResinType:
      type: string
      enum: [PP, PE, PS, ABS, PVC, PET, PC, PA6, PA66, EVA, PMMA, HDPE, LDPE, LLDPE, GPPS, HIPS, POM, EPDM, PEI, PETG, AS, MS, PVDC, Other]
    PPType:
      type: string
      enum: [ホモ, ブロック, ランダム]
    PEType:
      type: string
      enum: [LD, HD, LLD]
    PSType:
      type: string
      enum: [HI, GP]
    ABSType:
      type: string
      enum: ["難燃"]
    QuantityType:
      type: string
      enum: ["月間", "スポット"]
    PackagingType:
      type: string
      enum: ["紙袋", "フレコン", "カートン", "鉄箱", "ポリ袋"]
    ResinEntry:
      type: object
      properties:
        id:
          type: integer
        entryType:
          type: string
          enum: [source, demand]
        resinCategory:
          $ref: "#/components/schemas/ResinCategory"
        date:
          type: string
          format: date
        counterparty:
          type: string
        personInCharge:
          type: string
        resinType:
          $ref: "#/components/schemas/ResinType"
        manufacturer:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
        otherResinType:
          type: string
          nullable: true
        ppType:
          $ref: "#/components/schemas/PPType"
          nullable: true
        peType:
          $ref: "#/components/schemas/PEType"
          nullable: true
        psType:
          $ref: "#/components/schemas/PSType"
          nullable: true
        absType:
          $ref: "#/components/schemas/ABSType"
          nullable: true
        isClosed:
          type: string
        sampleAvailable:
          type: string
          nullable: true
        packaging:
          $ref: "#/components/schemas/PackagingType"
          nullable: true
        meltFlowIndexLower:
          type: number
          nullable: true
        meltFlowIndexUpper:
          type: number
          nullable: true
        charpyLower:
          type: number
          nullable: true
        charpyUpper:
          type: number
          nullable: true
        izodLower:
          type: number
          nullable: true
        izodUpper:
          type: number
          nullable: true
        densityLower:
          type: number
          nullable: true
        densityUpper:
          type: number
          nullable: true
        price:
          type: number
          nullable: true
        priceLower:
          type: number
          nullable: true
        priceUpper:
          type: number
          nullable: true
        locationType:
          type: string
          nullable: true
        storageLocation:
          type: string
          nullable: true
        quantity:
          type: number
          nullable: true
        quantityLower:
          type: number
          nullable: true
        quantityUpper:
          type: number
          nullable: true
        quantityType:
          $ref: "#/components/schemas/QuantityType"
          nullable: true
        remarks:
          type: string
          nullable: true
        prospectiveBuyer:
          type: string
          nullable: true
        desiredQuantity:
          type: number
          nullable: true
        proposedTo:
          type: string
          nullable: true
        sellingPrice:
          type: number
          nullable: true
        packagingWeight:
          type: number
          nullable: true
        plainMaker:
          type: string
          nullable: true
        usageType:
          type: string
          nullable: true
        finalNegotiatedPrice:
          type: number
          nullable: true
        origin:
          type: string
          nullable: true
        colorTone:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
          nullable: true
        tdsUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - entryType
        - resinCategory
        - date
        - counterparty
        - personInCharge
        - resinType
        - createdAt
        - updatedAt
    CreateResinEntry:
      type: object
      properties:
        entryType:
          type: string
          enum: [source, demand]
        resinCategory:
          $ref: "#/components/schemas/ResinCategory"
        date:
          type: string
          format: date
        counterparty:
          type: string
        personInCharge:
          type: string
        resinType:
          $ref: "#/components/schemas/ResinType"
        manufacturer:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
        otherResinType:
          type: string
          nullable: true
        ppType:
          $ref: "#/components/schemas/PPType"
          nullable: true
        peType:
          $ref: "#/components/schemas/PEType"
          nullable: true
        psType:
          $ref: "#/components/schemas/PSType"
          nullable: true
        absType:
          $ref: "#/components/schemas/ABSType"
          nullable: true
        isClosed:
          type: string
        sampleAvailable:
          type: string
          nullable: true
        packaging:
          $ref: "#/components/schemas/PackagingType"
          nullable: true
        meltFlowIndexLower:
          type: number
          nullable: true
        meltFlowIndexUpper:
          type: number
          nullable: true
        charpyLower:
          type: number
          nullable: true
        charpyUpper:
          type: number
          nullable: true
        izodLower:
          type: number
          nullable: true
        izodUpper:
          type: number
          nullable: true
        densityLower:
          type: number
          nullable: true
        densityUpper:
          type: number
          nullable: true
        price:
          type: number
          nullable: true
        priceLower:
          type: number
          nullable: true
        priceUpper:
          type: number
          nullable: true
        locationType:
          type: string
          nullable: true
        storageLocation:
          type: string
          nullable: true
        quantity:
          type: number
          nullable: true
        quantityLower:
          type: number
          nullable: true
        quantityUpper:
          type: number
          nullable: true
        quantityType:
          $ref: "#/components/schemas/QuantityType"
          nullable: true
        remarks:
          type: string
          nullable: true
        prospectiveBuyer:
          type: string
          nullable: true
        desiredQuantity:
          type: number
          nullable: true
        proposedTo:
          type: string
          nullable: true
        sellingPrice:
          type: number
          nullable: true
        packagingWeight:
          type: number
          nullable: true
        plainMaker:
          type: string
          nullable: true
        usageType:
          type: string
          nullable: true
        finalNegotiatedPrice:
          type: number
          nullable: true
        origin:
          type: string
          nullable: true
        colorTone:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
          nullable: true
        tdsUrl:
          type: string
          nullable: true
      required:
        - entryType
        - resinCategory
        - date
        - counterparty
        - personInCharge
        - resinType
    Match:
      type: object
      properties:
        source:
          $ref: "#/components/schemas/ResinEntry"
        demand:
          $ref: "#/components/schemas/ResinEntry"
        score:
          type: integer
          description: Match score 0-100
        reasons:
          type: array
          items:
            type: string
      required:
        - source
        - demand
        - score
        - reasons
...